lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
ask-sdk-model==1.23.0
paho-mqtt